fix list after update to solidStart@1.16
