+++ packages/web/package.json
@@ -11,35 +11,35 @@
   "type": "module",
   "devDependencies": {
     "@app/shared": "workspace:*",
-    "@mdx-js/rollup": "^2.0.0",
+    "@mdx-js/rollup": "^2.1.5",
     "@types/dlv": "^1.1.2",
-    "@types/node": "^18.6.4",
-    "@typescript-eslint/eslint-plugin": "^5.32.0",
-    "@typescript-eslint/parser": "^5.32.0",
-    "@unocss/reset": "^0.45.12",
-    "@unocss/transformer-directives": "^0.45.12",
-    "eslint": "^8.21.0",
+    "@types/node": "^18.11.4",
+    "@typescript-eslint/eslint-plugin": "^5.40.1",
+    "@typescript-eslint/parser": "^5.40.1",
+    "@unocss/reset": "^0.45.30",
+    "@unocss/transformer-directives": "^0.45.30",
+    "eslint": "^8.26.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-tailwindcss": "^3.6.0",
+    "eslint-plugin-tailwindcss": "^3.6.2",
     "prettier": "^2.7.1",
-    "rollup": "^2.78.1",
-    "sass": "^1.54.3",
-    "solid-start-node": "0.1.0-alpha.99",
-    "solid-start-vercel": "^0.1.6",
-    "typescript": "^4.4.3",
-    "unocss": "^0.45.12",
-    "vite": "^3.0.4"
+    "rollup": "^3.2.3",
+    "sass": "^1.55.0",
+    "solid-start-node": "0.2.0",
+    "solid-start-vercel": "^0.2.0",
+    "typescript": "^4.8.4",
+    "unocss": "^0.45.30",
+    "vite": "^3.1.8"
   },
   "dependencies": {
     "@analytics/google-analytics": "1.0.3",
-    "@solidjs/meta": "^0.28.0",
-    "@solidjs/router": "^0.4.2",
+    "@solidjs/meta": "^0.28.1",
+    "@solidjs/router": "^0.5.0",
     "analytics": "^0.8.1",
-    "date-fns": "^2.29.1",
-    "solid-js": "1.4.7",
+    "date-fns": "^2.29.3",
+    "solid-js": "1.6.0",
     "solid-mdx": "^0.0.6",
-    "solid-start": "0.1.0-alpha.99",
-    "undici": "^5.8.0"
+    "solid-start": "0.2.0",
+    "undici": "^5.11.0"
   }
 }
