+++ packages/web/src/routes/list/index.tsx
@@ -1,11 +1,11 @@
 import './list.scss'
+import { format } from 'date-fns'
 import { debounce } from '@app/shared'
 import { MetaSeo } from '~/components'
-import { Head, Title } from 'solid-start'
 import { createStore } from 'solid-js/store'
-import { ShariahFetcher, PageProps } from './fetcher'
+import { Head, Title, useRouteData } from 'solid-start'
 import { IMGS, pipe, TArrayConcat, TFilter, useTrackOnLoad } from '~/util'
-import { createEffect, createMemo, createResource, For, JSX, Show } from 'solid-js'
+import { createMemo, createResource, For, JSX, onMount, Show, Suspense } from 'solid-js'
 
 const staticExchangeColors = [
   'text-red-600 bg-red-100 border-red-500',
@@ -23,43 +23,97 @@ const props = {
   path: 'list',
 }
 
+export type PageProps = {
+  data: [exchange: string, code: string, name: string][]
+  metadata: Record<string, Date>
+  exchangesList: string[]
+  queryParams: Partial<{
+    q: string
+    exchange: string | string[]
+  }>
+}
+
+export function routeData({ location }) {
+  return createResource<PageProps | undefined>(
+    async () => {
+      const res = await fetch(import.meta.env.VITE_FETCH_URL + '/stock-list-human.json')
+      const { data, metadata } = await res.json()
+
+      const m = Object.entries(metadata).reduce(
+        (acc, [exchange, date]) => ({
+          ...acc,
+          [exchange]: format(new Date(date as Date), 'dd LLL yy'),
+        }),
+        {}
+      )
+
+      const search = new URLSearchParams(location.search) // { exchange: ['MYX', 'NASDAQ', 'SSE'], q: "search key" }
+
+      const queryParams = {
+        q: search.get('q') ?? '',
+        exchange: search.getAll('exchange') ?? [],
+      }
+
+      return {
+        data,
+        metadata: m,
+        queryParams,
+        exchangesList: queryParams.exchange.length === 0 ? Object.keys(metadata) : queryParams.exchange,
+      }
+    },
+    { initialValue: undefined }
+  )
+}
+
+const [store, setStore] = createStore({
+  data: [],
+  originalData: [],
+  exchangeStyle: {},
+  runFilter: () => {
+    setStore('data', transducerFilter())
+  },
+  metadata: {},
+  search: '',
+  currentExchange: {
+    _data: [], // exchangesList,
+    add(exchange: string) {
+      const s = new Set<string>(this._data)
+      s.add(exchange)
+      setStore('currentExchange', '_data', [...s])
+    },
+    delete(exchange: string) {
+      setStore(
+        'currentExchange',
+        '_data',
+        this._data.filter((ex) => ex !== exchange)
+      )
+    },
+    has(exchange: string): boolean {
+      return this._data.includes(exchange)
+    },
+  },
+})
+
+const showSearch = (item: PageProps['data']['0']): boolean => {
+  if (!store.search) {
+    return true
+  }
+
+  const searchRegex = new RegExp(store.search, 'ig')
+  return item.some((i) => searchRegex.test(i))
+}
+
+const showExchange = ([exchange]: PageProps['data']['0']): boolean => store.currentExchange.has(exchange)
+
+const transducerFilter = createMemo(() =>
+  store.originalData.reduce(pipe(TFilter(showExchange), TFilter(showSearch))(TArrayConcat), [])
+)
+
 export default function List(): JSX.Element {
   useTrackOnLoad()
+  const [data] = useRouteData<typeof routeData>()
 
-  const [data] = createResource<PageProps | undefined>(ShariahFetcher, {
-    initialValue: undefined,
-  })
-
-  const [store, setStore] = createStore({
-    data: [],
-    originalData: [],
-    exchangeStyle: {},
-    runFilter: () => {
-      setStore('data', transducerFilter())
-    },
-    metadata: {},
-    search: '',
-    currentExchange: {
-      _data: [], // exchangesList,
-      add(exchange: string) {
-        const s = new Set<string>(this._data)
-        s.add(exchange)
-        setStore('currentExchange', '_data', [...s])
-      },
-      delete(exchange: string) {
-        setStore(
-          'currentExchange',
-          '_data',
-          this._data.filter((ex) => ex !== exchange)
-        )
-      },
-      has(exchange: string): boolean {
-        return this._data.includes(exchange)
-      },
-    },
-  })
-
-  createEffect(() => {
+  onMount(() => {
     if (!data.loading) {
       const { data: originalData, metadata, exchangesList, queryParams } = data()
       setStore('data', originalData)
@@ -71,30 +125,12 @@ export default function List(): JSX.Element {
       setStore(
         'exchangeStyle',
         Object.keys(store.metadata).reduce(
-          (acc, key, i) => ({
-            ...acc,
-            [key]: staticExchangeColors[i % staticExchangeColors.length],
-          }),
+          (acc, key, i) => ({ ...acc, [key]: staticExchangeColors[i % staticExchangeColors.length] }),
           {}
         )
       )
     }
   })
-
-  const showSearch = (item: PageProps['data']['0']): boolean => {
-    if (!store.search) {
-      return true
-    }
-
-    const searchRegex = new RegExp(store.search, 'ig')
-    return item.some((i) => searchRegex.test(i))
-  }
-
-  const showExchange = ([exchange]: PageProps['data']['0']): boolean => store.currentExchange.has(exchange)
-
-  const transducerFilter = createMemo(() =>
-    store.originalData.reduce(pipe(TFilter(showExchange), TFilter(showSearch))(TArrayConcat), [])
-  )
 
   const handleSearchChange = debounce((e) => {
     const keyword = e.target.value
@@ -146,73 +182,78 @@ export default function List(): JSX.Element {
         <MetaSeo {...props} />
       </Head>
 
-      <Show
-        when={store.originalData.length}
+      <Suspense
         fallback={
           <div class='flex flex-col gap-2 justify-center items-center h-full text-gray-500'>
             <img src={IMGS.logo} alt='logo' class='animate-bounce' width='25px' height='25px' />
             Loading
           </div>
         }>
-        <div class='mx-auto max-w-3xl py-10 px-5'>
-          <svg class='hidden'>
-            <symbol
-              id='link-icon'
-              stroke='currentColor'
-              stroke-width='0'
-              class='text-gray-300 hover:text-green-200 fill-current'
-              viewBox='0 0 1024 1024'>
-              <path d='M574 665.4a8.03 8.03 0 0 0-11.3 0L446.5 781.6c-53.8 53.8-144.6 59.5-204 0-59.5-59.5-53.8-150.2 0-204l116.2-116.2c3.1-3.1 3.1-8.2 0-11.3l-39.8-39.8a8.03 8.03 0 0 0-11.3 0L191.4 526.5c-84.6 84.6-84.6 221.5 0 306s221.5 84.6 306 0l116.2-116.2c3.1-3.1 3.1-8.2 0-11.3L574 665.4zm258.6-474c-84.6-84.6-221.5-84.6-306 0L410.3 307.6a8.03 8.03 0 0 0 0 11.3l39.7 39.7c3.1 3.1 8.2 3.1 11.3 0l116.2-116.2c53.8-53.8 144.6-59.5 204 0 59.5 59.5 53.8 150.2 0 204L665.3 562.6a8.03 8.03 0 0 0 0 11.3l39.8 39.8c3.1 3.1 8.2 3.1 11.3 0l116.2-116.2c84.5-84.6 84.5-221.5 0-306.1zM610.1 372.3a8.03 8.03 0 0 0-11.3 0L372.3 598.7a8.03 8.03 0 0 0 0 11.3l39.6 39.6c3.1 3.1 8.2 3.1 11.3 0l226.4-226.4c3.1-3.1 3.1-8.2 0-11.3l-39.5-39.6z' />
-            </symbol>
-          </svg>
-          <input
-            type='text'
-            placeholder='Tesla, TOPGLOV, MYX, SZSE, NYSE, MYX, intel'
-            value={store.search}
-            class='px-2 mb-5 w-full h-10 rounded'
-            onInput={handleSearchChange}
-          />
-          <div class='flex gap-2 mb-4'>
-            <For each={Object.entries(store.metadata)}>
-              {([exchange]) => (
-                <label class='exchange'>
-                  <input
-                    type='checkbox'
-                    class='hidden'
-                    checked={store.currentExchange.has(exchange)}
-                    onClick={handleExchangeChange(exchange)}
-                  />
-                  <span>{exchange}</span>
-                </label>
-              )}
-            </For>
+        <Show when={data()} keyed>
+          <div class='mx-auto max-w-3xl py-10 px-5'>
+            <svg class='hidden'>
+              <symbol
+                id='link-icon'
+                stroke='currentColor'
+                stroke-width='0'
+                class='text-gray-300 hover:text-green-200 fill-current'
+                viewBox='0 0 1024 1024'>
+                <path d='M574 665.4a8.03 8.03 0 0 0-11.3 0L446.5 781.6c-53.8 53.8-144.6 59.5-204 0-59.5-59.5-53.8-150.2 0-204l116.2-116.2c3.1-3.1 3.1-8.2 0-11.3l-39.8-39.8a8.03 8.03 0 0 0-11.3 0L191.4 526.5c-84.6 84.6-84.6 221.5 0 306s221.5 84.6 306 0l116.2-116.2c3.1-3.1 3.1-8.2 0-11.3L574 665.4zm258.6-474c-84.6-84.6-221.5-84.6-306 0L410.3 307.6a8.03 8.03 0 0 0 0 11.3l39.7 39.7c3.1 3.1 8.2 3.1 11.3 0l116.2-116.2c53.8-53.8 144.6-59.5 204 0 59.5 59.5 53.8 150.2 0 204L665.3 562.6a8.03 8.03 0 0 0 0 11.3l39.8 39.8c3.1 3.1 8.2 3.1 11.3 0l116.2-116.2c84.5-84.6 84.5-221.5 0-306.1zM610.1 372.3a8.03 8.03 0 0 0-11.3 0L372.3 598.7a8.03 8.03 0 0 0 0 11.3l39.6 39.6c3.1 3.1 8.2 3.1 11.3 0l226.4-226.4c3.1-3.1 3.1-8.2 0-11.3l-39.5-39.6z' />
+              </symbol>
+            </svg>
+            <input
+              type='text'
+              placeholder='Tesla, TOPGLOV, MYX, SZSE, NYSE, MYX, intel'
+              value={store.search}
+              class='px-2 mb-5 w-full h-10 rounded'
+              onInput={handleSearchChange}
+            />
+            <div class='flex gap-2 mb-4'>
+              <For each={Object.entries(store.metadata)}>
+                {([exchange]) => (
+                  <label class='exchange'>
+                    <input
+                      type='checkbox'
+                      class='hidden'
+                      checked={store.currentExchange.has(exchange)}
+                      onClick={handleExchangeChange(exchange)}
+                    />
+                    <span>{exchange}</span>
+                  </label>
+                )}
+              </For>
+
+              <Show
+                keyed
+                when={store.search || store.currentExchange._data.length < Object.keys(store.metadata).length}
+                fallback={<div class='ml-auto' />}>
+                <button class='ml-auto text-green-500' onClick={handleClear}>
+                  Clear All
+                </button>
+              </Show>
+            </div>
 
             <Show
-              when={store.search || store.currentExchange._data.length < Object.keys(store.metadata).length}
-              fallback={<div class='ml-auto' />}>
-              <button class='ml-auto text-green-500' onClick={handleClear}>
-                Clear All
-              </button>
+              keyed
+              when={store.data.length}
+              fallback={
+                <div class='flex justify-center text-xl text-gray-500'>Please search or select new filter</div>
+              }>
+              <ShariahList data={transducerFilter()} exchangeStyle={store.exchangeStyle} />
             </Show>
           </div>
-
-          <Show
-            when={store.data.length}
-            fallback={<div class='flex justify-center text-xl text-white'>Please search or select new filter</div>}>
-            <ShariahList data={transducerFilter()} exchangeStyle={store.exchangeStyle} />
-          </Show>
-        </div>
-        <div class='flex flex-wrap gap-4 justify-center py-5'>
-          <For each={Object.entries(store.metadata)}>
-            {([exchange, date]) => (
-              <div class='py-1 px-2 bg-green-100 rounded'>
-                <span class='mr-1 text-xs text-green-800'>{exchange}</span>
-                <span class='font-bold text-green-900'>{date as unknown as string}</span>
-              </div>
-            )}
-          </For>
-        </div>
-      </Show>
+          <div class='flex flex-wrap gap-4 justify-center py-5'>
+            <For each={Object.entries(store.metadata)}>
+              {([exchange, date]) => (
+                <div class='py-1 px-2 bg-green-100 rounded'>
+                  <span class='mr-1 text-xs text-green-800'>{exchange}</span>
+                  <span class='font-bold text-green-900'>{date as unknown as string}</span>
+                </div>
+              )}
+            </For>
+          </div>
+        </Show>
+      </Suspense>
     </>
   )
 }
@@ -243,7 +284,7 @@ const ShariahListItem = (props: { item: PageProps['data']['0']; exchangeColor: s
       </a>
 
       {exchange && (
-        <Show when={exchange}>
+        <Show when={exchange} keyed>
           <span
             classList={{ [props.exchangeColor]: true }}
             class='hidden sm:block text-sm py-1 px-2 border rounded ml-auto opacity-80 group-hover:opacity-100'>
@@ -254,8 +295,3 @@ const ShariahListItem = (props: { item: PageProps['data']['0']; exchangeColor: s
     </li>
   )
 }
-
-// export const documentProps = {
-//   title: 'TSI: Shariah List',
-//   description: 'List of Shariah in Tradingview Shariah Indicator(TSI)',
-// }
